+++ src/pages/user/MyPage.jsx
@@ -1,6 +1,8 @@
+// MyPage.jsx
 import { Link } from "react-router-dom";
 import { useState, useEffect } from "react";
 import axios from "axios";
+import { WishPreview } from "./components/WishPreview";
 
 function MyPage() {
   const userId = 1;
@@ -24,10 +26,6 @@ function MyPage() {
       5: 0,
     }
   });
-
-  const [wishlistItems, setWishlistItems] = useState([]);
-  const [imageErrors, setImageErrors] = useState({});
-  const [isWishlistLoading, setIsWishlistLoading] = useState(true);
 
   const loadDashboardInfo = async () => {
     try {
@@ -60,39 +58,9 @@ function MyPage() {
     }
   }
 
-  const loadWishlistItems = async () => {
-    setIsWishlistLoading(true);
-    try {
-      const response = await axios.get(
-        `http://localhost:8080/api/my/user/dashboard/preview`,
-        { params: { userId } }
-      );
-
-      const limitedData = response.data.data.slice(0, 4);
-      setWishlistItems(limitedData)
-
-    } catch (error) {
-      console.error("찜 리스트를 불러올 수 없습니다:", error.message);
-    } finally {
-      setIsWishlistLoading(false);
-    }
-  };
-
   useEffect(() => {
     loadDashboardInfo();
-    loadWishlistItems();
   }, []);
-
-  const formatPrice = (price) => {
-    return price.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ',');
-  };
-
-  const formatImageUrl = (imagePath) => {
-    const baseUrl = "https://u7ouobpu9909.edge.naverncp.com/cdie6Z8lNS/wishlist/";
-    const params = "?type=f&w=216&h=180"
-
-    return imagePath.startsWith('http') ? imagePath : `${baseUrl}${imagePath}${params}`;
-  };
 
   return (
     <div className="w-full m-7">
@@ -196,70 +164,7 @@ function MyPage() {
           </div>
         </div>
       </div>
-
-      <div className="flex flex-col mt-12 mx-8 mb-8 select-none">
-        <div className="flex justify-between items-end border-b-2 pb-2 border-gray-dark">
-          <h2 className="font-bold text-2xl text-brown">찜 리스트</h2>
-          <Link to="/my/user/wishlist" className="text-text-gray text-xs font-bold hover:text-brown transition-colors">
-            더보기＞
-          </Link>
-        </div>
-        
-        {isWishlistLoading ? (
-          <div className="flex w-full justify-center mt-6 h-[200px] items-center">
-            <div className="animate-spin rounded-full h-12 w-12 border-t-2 border-b-2 border-brown"></div>
-          </div>
-        ) : wishlistItems.length > 0 ? (
-          <div className="grid grid-cols-4 gap-4 mt-6">
-            {wishlistItems.map((item) => (
-              // 상품 클릭 시 상품 상세 페이지로 이동
-              <Link 
-                to={`/product/${item.saleId}`} 
-                key={item.id}
-                className="bg-white border border-gray-dark rounded-lg overflow-hidden shadow-md hover:shadow-lg transition-shadow"
-              >
-                <div className="relative h-48 overflow-hidden">
-                  <div className="w-full h-full bg-gray-light flex items-center justify-center">
-                    {item.saveFile && !imageErrors[item.id] ? (
-                      <img 
-                        src={formatImageUrl(item.saveFile)}
-                        alt={item.title} 
-                        className="w-full h-full object-cover"
-                        onError={() => {
-                          setImageErrors((prev) => ({ ...prev, [item.id]: true }));
-                        }}
-                      />
-                    ) : (
-                      <svg xmlns="http://www.w3.org/2000/svg" className="h-16 w-16 text-gray-dark" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                        <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1} d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z" />
-                      </svg>
-                    )}
-                  </div>
-                </div>
-                
-                <div className="p-4">
-                  <h3 className="font-bold text-brown truncate">{item.title}</h3>
-                  <p className="text-text-gray text-sm mt-1">{item.name}</p>
-                  <div className="mt-2">
-                    <span className="text-brown font-bold">{formatPrice(item.price)}원</span>
-                  </div>
-                </div>
-              </Link>
-            ))}
-          </div>
-        ) : (
-          <div className="flex flex-col w-full justify-center items-center mt-6 h-[200px]">
-            <svg xmlns="http://www.w3.org/2000/svg" className="h-12 w-12 text-text-gray mb-2" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={1.5} d="M4.318 6.318a4.5 4.5 0 000 6.364L12 20.364l7.682-7.682a4.5 4.5 0 00-6.364-6.364L12 7.636l-1.318-1.318a4.5 4.5 0 00-6.364 0z" />
-            </svg>
-            <span className="text-text-gray text-lg">찜한 상품이 없습니다.</span>
-            {/* 실시간 랭킹 상품 목록 url로 변경 */}
-            <Link to="/product/list" className="mt-4 px-4 py-2 bg-brown text-white rounded-md hover:bg-brown-dark transition-colors text-sm">
-              상품 둘러보기
-            </Link>
-          </div>
-        )}
-      </div>
+      <WishPreview userId={userId} />
     </div>
   );
 }
